+++ execution/evm/abi/abi.go
@@ -14,7 +14,7 @@ import (
 
 	burrow_binary "github.com/hyperledger/burrow/binary"
 	"github.com/hyperledger/burrow/crypto"
-	"github.com/hyperledger/burrow/execution/evm/sha3"
+	"github.com/hyperledger/burrow/crypto/sha3"
 )
 
 // EVM Solidity calls and return values are packed into
@@ -1098,7 +1098,7 @@ func UnpackRevert(data []byte) (message *string, err error) {
 /*
  * Given a eventSpec, get all the fields (topic fields or not)
  */
-func UnpackEvent(eventSpec EventSpec, topics []burrow_binary.Word256, data []byte, args ...interface{}) error {
+func UnpackEvent(eventSpec *EventSpec, topics []burrow_binary.Word256, data []byte, args ...interface{}) error {
 	// First unpack the topic fields
 	topicIndex := 0
 	if !eventSpec.Anonymous {