+++ CHANGELOG.md
@@ -1,4 +1,19 @@
 # [Hyperledger Burrow](https://github.com/hyperledger/burrow) Changelog
+## [Unreleased]
+### Security
+
+### Changed
+
+### Fixed
+
+### Added
+- [Deploy] burrow deploy now prints events generated during transactions
+
+### Removed
+
+### Deprecated
+
+
 ## [0.23.1] - 2018-11-14
 ### Fixed
 - [EVM] state/Cache no longer allows SetStorage on accounts that do not exist
@@ -321,6 +336,7 @@ This release marks the start of Eris-DB as the full permissioned blockchain node
   - [Blockchain] Fix getBlocks to respect block height cap.
 
 
+[Unreleased]: https://github.com/hyperledger/burrow/compare/v0.23.1...HEAD
 [0.23.1]: https://github.com/hyperledger/burrow/compare/v0.23.0...v0.23.1
 [0.23.0]: https://github.com/hyperledger/burrow/compare/v0.22.0...v0.23.0
 [0.22.0]: https://github.com/hyperledger/burrow/compare/v0.21.0...v0.22.0